# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'spring-boot-api-example'","description":"'Complete example to develop an api in Spring Boot'","contact":{"name":"Hernán Jesús Morais","email":"hmorais@frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Complete example to develop an api in Spring Boot'"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"Create a new player","description":"Return the player created with your id. If a player exist with the username or email, then return 404. Additionally, the email must be valid and the password must have at least 8 characters and contain at least one number, one lowercase letter, one uppercase letter and one special character.","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a player by id","description":"Return a player by they id. If the player doesn't exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"The Player id","format":"int64","example":1},"userName":{"title":"Player user name","type":"string","description":"The Player user name","example":"myUserName"},"password":{"title":"Player password","type":"string","description":"The Player password","example":"Password03#"},"email":{"title":"Player email","type":"string","description":"The Player email","example":"email@email.com"},"avatar":{"title":"Player avatar url","type":"string","description":"The Player avatar","nullable":true,"example":"https://localhost:8080/avatars/myUserName"},"lastLoginDate":{"title":"Player last login","type":"string","description":"The Player last login","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}